+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Add `validOnSelect`, `disabled` props to the `TextInputAndSelect` component.
+- Fix: Add `valid`, `disabled` props to the `SelectWithState` component.
+- Fix: Add `checked-disabled` and `arrow-bottom-disabled` svgs.
+  Please update `@import k-Select((…));` to `@import k-Select;`.
+- Fix: Update disabled color on `TextInput`, `TextInputAndSelect`, `Button`
+  and `Select`.
+- Breaking change: The k-Select mixin does not accept options anymore.
 - Breaking change: Remove `react-on-rails` module as dependency.
   To be retro-compatible, add the dependency on your project's `package.json`:
   ```